# backend/docs/swagger.yaml
basePath: /api
consumes:
- application/json
definitions:
...
  github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.ChangeUserPassword:
    properties:
      password:
        maxLength: 16
        minLength: 4
        type: string
    required:
    - password
    type: object
  github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.ChangeUserPasswordBySelf:
    properties:
      new_password:
        maxLength: 16
        minLength: 4
        type: string
      old_password:
        maxLength: 16
        minLength: 4
        type: string
    required:
    - new_password
    - old_password
    type: object
  github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.CreateUserData:
    properties:
      password:
        maxLength: 16
        minLength: 4
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.DashboardRelease:
    properties:
      current:
        type: string
      latest:
        type: string
    required:
    - current
    - latest
    type: object
  github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.GetSystemInitResponse:
    properties:
      first_init:
        type: boolean
      google_captcha_site_key:
        type: string
      telegram_bot_enabled:
        type: boolean
    type: object
  github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.GetSystemResponse:
    properties:
      auto_delete_inactive_users:
        type: boolean
      client_profile_connection_name:
        type: string
      client_profile_server_address:
        type: string
      client_profile_server_port:
        type: integer
      first_init:
        type: boolean
      google_captcha_secret_key:
        type: string
      google_captcha_site_key:
        type: string
      keep_inactive_user_days:
        type: integer
    type: object
  github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.LoginData:
    properties:
      password:
        example: doe123456
        maxLength: 16
        minLength: 2
        type: string
      remember_me:
        type: boolean
      token:
        type: string
      username:
        example: john_doe
        maxLength: 16
        minLength: 2
        type: string
    required:
    - password
    - username
    type: object
  github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.PatchSystemUpdateData:
    properties:
      auto_delete_inactive_users:
        type: boolean
      client_profile_connection_name:
        type: string
      client_profile_server_address:
        type: string
      client_profile_server_port:
        type: integer
      google_captcha_secret_key:
        type: string
      google_captcha_site_key:
        type: string
      keep_inactive_user_days:
        type: integer
    required:
    - auto_delete_inactive_users
    - client_profile_connection_name
    - client_profile_server_address
    - client_profile_server_port
    - google_captcha_secret_key
    - google_captcha_site_key
    - keep_inactive_user_days
    type: object
  github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.ResetAdminPassword:
    properties:
      new_password:
        maxLength: 16
        minLength: 4
        type: string
      secret_key:
        maxLength: 64
        minLength: 16
        type: string
    required:
    - new_password
    - secret_key
    type: object
  github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.ResetPasswordResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    required:
    - token
    - user
    type: object
  github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.UserLoginResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/models.User'
    required:
    - token
    - user
    type: object
  github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.UsersResponse:
    properties:
      meta:
        $ref: '#/definitions/request.Meta'
      result:
        items:
          $ref: '#/definitions/models.User'
        type: array
    required:
    - meta
    type: object
...
  middlewares.Unauthorized:
    properties:
      error:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      last_login:
        type: string
      superadmin:
        type: boolean
      updated_at:
        type: string
      username:
        type: string
    required:
    - id
    - last_login
    - superadmin
    - username
    type: object
  models.UsersLookup:
    properties:
      id:
        type: integer
      username:
        type: string
    required:
    - id
    - username
    type: object
...
  request.ErrorResponse:
    properties:
      error:
        items:
          type: string
        type: array
      message:
        items:
          type: string
        type: array
    required:
    - error
    - message
    type: object
  request.Meta:
    properties:
      page:
        type: integer
      size:
        type: integer
      total_records:
        type: integer
    required:
    - page
    - size
    - total_records
    type: object
...
info:
  contact: {}
  description: Administrative API for managing Ocserv Dashboard users, groups, reports,
    runtime, backups, system settings, Telegram integrations, and agent nodes.
  title: Ocserv Dashboard Admin API
  version: "1.0"
paths:
...
  /system:
    get:
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.GetSystemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Get panel System Config
      tags:
      - System
    patch:
      consumes:
      - application/json
      parameters:
      - description: update system config data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.PatchSystemUpdateData'
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.GetSystemResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Update panel System Config
      tags:
      - System
  /system/init:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.GetSystemInitResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Get panel System init Config
      tags:
      - System
...
  /system/release:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.DashboardRelease'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Get Dashboard the current and latest release
      tags:
      - System
  /system/user/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset admin password data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.ResetAdminPassword'
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.ResetPasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Reset admin password by secret key
      tags:
      - System(User)
  /system/users:
    get:
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.UsersResponse'
      summary: List of Admin or simple users
      tags:
      - System(Users)
    post:
      consumes:
      - application/json
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: create user data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.CreateUserData'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Create user
      tags:
      - System(Users)
  /system/users/{id}:
    delete:
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete simple user
      tags:
      - System(Users)
  /system/users/{id}/password:
    post:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: user new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.ChangeUserPassword'
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Change user password by admin
      tags:
      - System(Users)
  /system/users/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: login data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.LoginData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.UserLoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
      summary: Admin users login
      tags:
      - System(Users)
  /system/users/lookup:
    get:
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UsersLookup'
            type: array
      summary: List of Users Lookup
      tags:
      - System(Users)
  /system/users/password:
    post:
      consumes:
      - application/json
      parameters:
      - description: user new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_mmtaee_ocserv-dashboard_backend_internal_services_admin_api_system.ChangeUserPasswordBySelf'
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Change user password by self
      tags:
      - System(Users)
  /system/users/profile:
    get:
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      summary: Get User Profile
      tags:
      - System(Users)
...
produces:
- application/json
swagger: "2.0"